Logging docs (and moved some logging classes), security notes and threading notes
