+++ src/main/java/io/vertx/core/file/impl/AsyncFileImpl.java
@@ -28,7 +28,7 @@ import io.vertx.core.impl.Arguments;
 import io.vertx.core.impl.ContextImpl;
 import io.vertx.core.impl.VertxInternal;
 import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.impl.LoggerFactory;
+import io.vertx.core.logging.LoggerFactory;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
