+++ src/main/java/io/vertx/core/file/impl/WindowsFileSystem.java
@@ -23,7 +23,7 @@ import io.vertx.core.file.OpenOptions;
 import io.vertx.core.impl.ContextImpl;
 import io.vertx.core.impl.VertxInternal;
 import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.impl.LoggerFactory;
+import io.vertx.core.logging.LoggerFactory;
 
 import java.util.Objects;
