+++ src/main/java/io/vertx/core/http/impl/HttpClientImpl.java
@@ -52,7 +52,7 @@ import io.vertx.core.impl.Closeable;
 import io.vertx.core.impl.ContextImpl;
 import io.vertx.core.impl.VertxInternal;
 import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.impl.LoggerFactory;
+import io.vertx.core.logging.LoggerFactory;
 import io.vertx.core.spi.metrics.Metrics;
 import io.vertx.core.spi.metrics.HttpClientMetrics;
 import io.vertx.core.net.impl.KeyStoreHelper;
