+++ src/main/java/io/vertx/core/http/impl/HttpServerResponseImpl.java
@@ -30,7 +30,7 @@ import io.vertx.core.http.HttpServerResponse;
 import io.vertx.core.impl.ContextImpl;
 import io.vertx.core.impl.VertxInternal;
 import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.impl.LoggerFactory;
+import io.vertx.core.logging.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
