+++ src/main/java/io/vertx/core/impl/OrderedExecutorFactory.java
@@ -18,7 +18,7 @@
 package io.vertx.core.impl;
 
 import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.impl.LoggerFactory;
+import io.vertx.core.logging.LoggerFactory;
 
 import java.util.LinkedList;
 import java.util.concurrent.Executor;
