+++ src/main/java/io/vertx/core/impl/Redeployer.java
@@ -17,7 +17,7 @@
 package io.vertx.core.impl;
 
 import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.impl.LoggerFactory;
+import io.vertx.core.logging.LoggerFactory;
 
 import java.io.File;
 import java.util.HashMap;
