+++ src/main/java/io/vertx/core/impl/verticle/CompilingClassLoader.java
@@ -17,7 +17,7 @@
 package io.vertx.core.impl.verticle;
 
 import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.impl.LoggerFactory;
+import io.vertx.core.logging.LoggerFactory;
 
 import javax.tools.Diagnostic;
 import javax.tools.DiagnosticCollector;
