+++ src/main/java/io/vertx/core/logging/Log4jLogDelegate.java
@@ -14,13 +14,14 @@
  * You may elect to redistribute this code under either of these licenses.
  */
 
-package io.vertx.core.logging.impl;
+package io.vertx.core.logging;
 
 import io.vertx.core.logging.Logger;
+import io.vertx.core.spi.logging.LogDelegate;
 import org.apache.log4j.Level;
 
 /**
- * A {@link LogDelegate} which delegates to Apache Log4j
+ * A {@link io.vertx.core.spi.logging.LogDelegate} which delegates to Apache Log4j
  *
  * @author <a href="kenny.macleod@kizoom.com">Kenny MacLeod</a>
  */
