+++ src/main/java/io/vertx/core/logging/Logger.java
@@ -16,7 +16,7 @@
 
 package io.vertx.core.logging;
 
-import io.vertx.core.logging.impl.LogDelegate;
+import io.vertx.core.spi.logging.LogDelegate;
 
 /**
  * This class allows us to isolate all our logging dependencies in one place. It also allows us to have zero runtime
@@ -29,7 +29,7 @@ import io.vertx.core.logging.impl.LogDelegate;
  * If you would prefer to use Log4J or SLF4J instead of JUL then you can set a system property called
  * {@code vertx.logger-delegate-factory-class-name} to the class name of the delegate for your logging system.
  * For Log4J the value is {@code io.vertx.core.logging.impl.Log4JLogDelegateFactory}, for SLF4J the value
- * is {@code io.vertx.core.logging.impl.SLF4JLogDelegateFactory}. You will need to ensure whatever jar files
+ * is {@code io.vertx.core.logging.SLF4JLogDelegateFactory}. You will need to ensure whatever jar files
  * required by your favourite log framework are on your classpath.
  *
  * @author <a href="mailto:tim.fox@jboss.com">Tim Fox</a>
