+++ src/main/java/io/vertx/core/logging/SLF4JLogDelegateFactory.java
@@ -14,7 +14,10 @@
  * You may elect to redistribute this code under either of these licenses.
  */
 
-package io.vertx.core.logging.impl;
+package io.vertx.core.logging;
+
+import io.vertx.core.spi.logging.LogDelegate;
+import io.vertx.core.spi.logging.LogDelegateFactory;
 
 /*
  * @author <a href="http://tfox.org">Tim Fox</a>
