+++ src/main/java/io/vertx/core/net/impl/SocketDefaults.java
@@ -17,7 +17,7 @@
 package io.vertx.core.net.impl;
 
 import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.impl.LoggerFactory;
+import io.vertx.core.logging.LoggerFactory;
 
 import java.net.Socket;
