+++ src/main/java/io/vertx/core/package-info.java
@@ -637,7 +637,15 @@
  *
  * == Thread safety
  *
- * Notes on thread safety of Vert.x objects
+ * Most Vert.x objects are safe to access from different threads. _However_ performance is optimised when they are
+ * accessed from the same context they were created from.
+ *
+ * For example if you have deployed a verticle which creates a {@link io.vertx.core.net.NetServer} which provides
+ * {@link io.vertx.core.net.NetSocket} instances in it's handler, then it's best to always access that socket instance
+ * from the event loop of the verticle.
+ *
+ * If you stick to the standard Vert.x verticle deployment model and avoid sharing objects between verticles then this
+ * should be the case without you having to think about it.
  *
  * == Metrics SPI
  *
@@ -879,19 +887,79 @@
  * You can also specify cluster managers programmatically if embedding Vert.x using
  * {@link io.vertx.core.VertxOptions#setClusterManager(io.vertx.core.spi.cluster.ClusterManager)}.
  *
+ * == Logging
+ *
+ * Vert.x logs using it's in-built logging API. The default implementation uses the JDK (JUL) logging so no extra
+ * logging dependencies are needed.
+ *
+ * === Configuring JUL logging
+ *
+ * A JUL logging configuration file can be specified in the normal JUL way by providing a system property called:
+ * `java.util.logging.config.file` with the value being your configuration file. For more information on this and
+ * the structure of a JUL config file please consult the JUL logging documentation.
+ *
+ * Vert.x also provides a slightly more convenient way to specify a configuration file without having to set a system
+ * property. Just provide a JUL config file with the name `default-jul-logging.properties` on your classpath (e.g.
+ * inside your fatjar) and Vert.x will use that to configure JUL.
+ *
+ * === Using another logging framework
+ *
+ * If you don't want Vert.x to use JUL for it's own logging you can configure it to use another logging framework, e.g.
+ * Log4J or SLF4J.
+ *
+ * To do this you should set a system property called `vertx.logger-delegate-factory-class-name` with the name of a Java
+ * class which implements the interface {@link io.vertx.core.logging.LoggerFactory}. We provide pre-built implementations for
+ * Log4J and SLF4J with the class names `io.vertx.core.logging.Log4jLogDelegateFactory` and `io.vertx.core.logging.SLF4JLogDelegateFactory`
+ * respectively. If you want to use these implementations you should also make sure the relevant Log4J or SLF4J jars
+ * are on your classpath.
+ *
+ * === Logging from your application
+ *
+ * Vert.x itself is just a library and you can use whatever logging library you prefer to log from your own application,
+ * using that logging library's API.
+ *
+ * However, if you prefer you can use the Vert.x logging facility as described above to provide logging for your
+ * application too.
+ *
+ * To do that you use {@link io.vertx.core.logging.LoggerFactory} to get an instance of {@link io.vertx.core.logging.Logger}
+ * which you then use for logging, e.g.
+ *
+ * [source,$lang]
+ * ----
+ * {@link examples.CoreExamples#example18}
+ * ----
  *
  * === High Availability
  *
  * == Security notes
  *
- * Warn about file uploads and serving files from arbitrary locations
+ * Vert.x is a toolkit, not an opinionated framework where we force you to do things in a certain way. This gives you
+ * great power as a developer but with that comes great responsibility.
  *
- * Vert.x is a tool kit
+ * As with any toolkit, it's possible to write insecure applications, so you should always be careful when developing
+ * your application especially if it's exposed to the public (e.g. over the internet).
  *
- * Run in a security sandbox
+ * === Web applications
  *
- * Use Apex
+ * If writing a web application it's highly recommended that you use Vert.x-Web instead of Vert.x core directly for
+ * serving resources and handling file uploads. Vert.x-Web normalises the path in requests to prevent malicious clients
+ * from crafting URLs to access resources outside of the web root. Similarly for file uploads Vert,x-Web provides
+ * functionality for uploading to a known place on disk and does not rely on the filename provided by the client in the
+ * upload which could be crafted to upload to a different place on disk. Vert.x core itself does not provide such checks
+ * so it would be up to you as a developer to implement them yourself.
  *
+ * === Clustered event bus traffic
+ *
+ * When clustering the event bus between different Vert.x nodes on a network, the traffic is sent un-encrypted across the
+ * wire, so do not use this if you have confidential data to send and your Vert.x nodes are not on a trusted network.
+ *
+ * === Standard security best practices
+ *
+ * Any service can have potentially vulnerabilities whether it's written using Vert.x or any other toolkit so always
+ * follow security best practice, especially if your service is public facing.
+ *
+ * For example you should always run them in a DMZ and with an user account that has limited rights in order to limit
+ * the extent of damage in case the service was compromised.
  *
  */
 @Document(fileName = "index.adoc")
