+++ src/test/resources/vertx-default-jul-logging.properties
@@ -18,7 +18,7 @@ java.util.logging.SimpleFormatter.format=%5$s %6$s\n
 java.util.logging.ConsoleHandler.formatter=java.util.logging.SimpleFormatter
 java.util.logging.ConsoleHandler.level=INFO
 java.util.logging.FileHandler.level=INFO
-java.util.logging.FileHandler.formatter=io.vertx.core.logging.impl.VertxLoggerFormatter
+java.util.logging.FileHandler.formatter=io.vertx.core.logging.VertxLoggerFormatter
 
 # Put the log in the system temporary directory
 java.util.logging.FileHandler.pattern=%t/vertx.log
